requirements.txt
python-telegram-bot==13.15
urllib3==1.26.15
python-dotenv==1.0.0
pillow>=10.0.0
pyngrok>=5.0.0
gunicorn==21.2.0
Flask==2.3.3
Werkzeug==2.3.7 